# src/generated/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"MedicalFacility": {
				"description": "Medical Facility / Cơ Sở Khám Chữa Bệnh (Code 063)\nHealthcare facilities for BHXH registration",
				"properties": {
					"ma": {
						"type": "string",
						"description": "Facility code"
					},
					"ten": {
						"type": "string",
						"description": "Facility name (Vietnamese)"
					},
					"matinh": {
						"type": "string",
						"description": "Province code"
					},
					"mahuyen": {
						"type": "string",
						"description": "District code"
					},
					"maphuongxa": {
						"type": "string",
						"description": "Ward code"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GeographicListResponse_MedicalFacility_": {
				"description": "Geographic list response wrapper",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/MedicalFacility"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Province": {
				"description": "Province/Tinh/Thanh Pho entity",
				"properties": {
					"ma": {
						"type": "string",
						"description": "Province code"
					},
					"ten": {
						"type": "string",
						"description": "Province name (Vietnamese)"
					}
				},
				"required": [
					"ma",
					"ten"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeographicListResponse_Province_": {
				"description": "Geographic list response wrapper",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Province"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Ward": {
				"description": "Ward/Phuong/Xa entity",
				"properties": {
					"ma": {
						"type": "string",
						"description": "Ward code"
					},
					"ten": {
						"type": "string",
						"description": "Ward name (Vietnamese)"
					}
				},
				"required": [
					"ma",
					"ten"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeographicListResponse_Ward_": {
				"description": "Geographic list response wrapper",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Ward"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "X-API-Key",
				"description": "API key for authentication",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "bhxh-api",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/api/v1/master-data/medical-facilities": {
			"get": {
				"operationId": "GetMedicalFacilities",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeographicListResponse_MedicalFacility_"
								}
							}
						}
					},
					"500": {
						"description": "Failed to fetch medical facilities",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get medical facilities (Cơ Sở Khám Chữa Bệnh - Code 063)\nPublic endpoint - no authentication required\nHealthcare facilities for BHXH registration: hospitals, clinics, medical stations",
				"security": [],
				"parameters": [
					{
						"description": "Province code (2-digit, e.g., \"79\" for HCMC). If omitted, returns all facilities.",
						"in": "query",
						"name": "maTinh",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/geographic/provinces": {
			"get": {
				"operationId": "GetProvinces",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeographicListResponse_Province_"
								}
							}
						}
					},
					"500": {
						"description": "Failed to fetch provinces",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get all provinces (Tinh/Thanh pho)\nPublic endpoint - no authentication required",
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/geographic/wards": {
			"get": {
				"operationId": "GetWards",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeographicListResponse_Ward_"
								}
							}
						}
					},
					"400": {
						"description": "Missing matinh parameter",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Failed to fetch wards",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get wards by province and district\nPublic endpoint - no authentication required",
				"security": [],
				"parameters": [
					{
						"description": "Province code (2-digit, required)",
						"in": "query",
						"name": "matinh",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "District code (optional)",
						"in": "query",
						"name": "mahuyen",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/geographic/districts": {
			"get": {
				"operationId": "GetDistricts",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeographicListResponse_MedicalFacility_"
								}
							}
						}
					},
					"500": {
						"description": "Failed to fetch medical facilities",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "Province code (2-digit, e.g., \"79\" for Ho Chi Minh City). If omitted, returns all.",
						"in": "query",
						"name": "maTinh",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}